// src/core/types.d.ts
export type ValueType = 'boolean' | 'integer' | 'float' | 'string' | 'datetime' | 'date' | 'time';
export type DataType = 'varchar' | 'text' | 'json' | 'int' | 'tinyint' | 'smallint' | 'double' | 'float' | 'decimal' | 'datetime' | 'date' | 'time';
export type GridColSizeType = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 'auto';
export type CharsetType = 'utf8';
export type ColorType = 'primary' | 'secondary' | 'tertiary' | 'success' | 'warning' | 'danger' | 'light' | 'medium' | 'dark';
export type SlotType = 'start' | 'end' | 'icon-only';
export type ModeType = 'ios' | 'md';
export type FormBehaviorType = 'touched' | 'dirty' | 'isNew';
export type stringOrExpr = string | {
    expr: string;
};
// ...
export interface FocusableControlType {
    /**
     * Emitted when the control loses focus.
     *
     * To obtain the event reference the argument must be named `$event`.
     * ```typescript
     * input: { ionBlur: `inputOnBlur($event)` }
     * ```
     * ```html
     * <ion-input (ionBlur)="inputOnBlur($event)">
     * ```
     *
     * @category Event
     */
    ionBlur?: string | undefined;
    /**
     * Emitted when the control has focus.
     *
     * To obtain the event reference the argument must be named `$event`.
     * ```typescript
     * input: { ionFocus: `inputFocused($event)` }
     * ```
     * ```html
     * <ion-input (ionFocus)="inputFocused($event)">
     * ```
     *
     * @category Event
     */
    ionFocus?: string | undefined;
}
// ...
export interface IconControlType extends BaseControlType {
    /** The color to use for the background of the item. */
    color?: ColorType | {
        expr: string;
    } | undefined;
    /** Specifies which icon to use on ios mode. */
    ios?: stringOrExpr;
    /** Specifies which icon to use on md mode. */
    md?: stringOrExpr;
    /** The mode determines which platform styles to use. */
    mode?: ModeType;
    /** Specifies which icon to use from the built-in set of icons.
     *
     * ```typescript
     * component: { icon: 'people' }
     * ```
     * <br>
     *
     * ```html
     * <ion-icon name="people">
     * ```
     */
    name?: stringOrExpr;
    /** The size of the icon. Available options are: "small" and "large".
     *
     * ```typescript
     * component: { icon: { size: 'small' } }
     * ```
     * <br>
     *
     * ```html
     * <ion-icon size="small">
     * ```
     */
    size?: 'small' | 'large';
    /** Specifies the exact src of an SVG file to use.
     *
     * ```typescript
     * component: { icon: 'people.svg' }
     * ```
     * <br>
     *
     * ```html
     * <ion-icon src="assets/icon/people.svg">
     * ```
     */
    src?: stringOrExpr;
    /** Content is placed between the named slots if provided without a slot.
     *
     * ```typescript
     * component: { icon: { slot: 'start' }
     * ```
     * <br>
     *
     * ```html
     * <ion-icon slot="start">
     * ```
     */
    slot?: SlotType;
    /** Expression evaluated to determine the visibility of the component.
     *
     * ```typescript
     * component: { icon: { visible: '!isNew' }
     * ```
     *
     * ```html
     * <ion-icon *ngIf="!isNew">
     * ```
     */
    visible?: boolean | string;
}
export interface LabelControlType extends BaseControlType {
    /** The color to use for the text of the label. */
    color?: ColorType | {
        expr: string;
    } | undefined;
    /** The mode determines which platform styles to use. */
    mode?: ModeType;
    /** The position determines where and how the label behaves inside an item. */
    position?: 'floating' | 'fixed' | 'stacked' | undefined;
    /** The text to show in the label.
     *
     * To evaluate the text as an expression, use the following notation:
     *
     * ```typescript
     * label: { text: { expr: `isNew ? 'buttons.cancel' : 'buttons.delete'` }},
     * ```
     */
    text?: stringOrExpr;
    /** Pipes to apply to the text.
     *
     * - If not specified, the translation pipe is applied by default.
     *
     * ```typescript
     * label: 'buttons.accept',
     * ```
     *
     * ```html
     * <ion-label>{{'buttons.accept' | translate }}</ion-label>
     * ```
     *
     * - Pipes are applied in the order they are declared.
     *
     * ```typescript
     * label: { text: 'buttons.accept', pipes: ['translate', 'uppercase'] },
     * ```
     *
     * ```html
     * <ion-label>{{'buttons.accept' | translate | uppercase }}</ion-label>
     * ```
     *
     * - To avoid the default translation pipe:
     *
     * ```typescript
     * label: { text: 'Aceptar', pipes: [] },
     * ```
     *
     * ```html
     * <ion-label>Aceptar</ion-label>
     * ```
     */
    pipes?: ('translate' | 'uppercase' | 'lowercase' | 'titlecase')[];
}
export interface ButtonControlType extends BaseControlType, FocusableControlType {
    /** The type of button. */
    buttonType?: 'button' | 'submit' | 'reset';
    /** The color to use from your application's color palette. Default options are: "primary", "secondary", "tertiary", "success", "warning", "danger", "light", "medium", and "dark". For more information on colors, see theming. */
    color?: ColorType | {
        expr: string;
    } | undefined;
    /** This attribute instructs browsers to download a URL instead of navigating to it, so the user will be prompted to save it as a local file. If the attribute has a value, it is used as the pre-filled file name in the Save prompt (the user can still change the file name if they want). */
    download?: string | undefined;
    /** Set to "block" for a full-width button or to "full" for a full-width button without left and right borders. */
    expand?: 'block' | 'full' | undefined;
    /** Set to "clear" for a transparent button, to "outline" for a transparent button with a border, or to "solid". The default style is "solid" except inside of a toolbar, where the default is "clear". */
    fill?: 'clear' | 'default' | 'outline' | 'solid' | undefined;
    /** Contains a URL or a URL fragment that the hyperlink points to. If this property is set, an anchor tag will be rendered. */
    href?: string | undefined;
    /** The mode determines which platform styles to use. */
    mode?: ModeType;
    /** Specifies the relationship of the target object to the link object. The value is a space-separated list of link types. */
    rel?: string | undefined;
    /** When using a router, it specifies the transition direction when navigating to another page using href. */
    routerDirection?: 'back' | 'forward' | 'root';
    /** The button shape. */
    shape?: 'round' | undefined;
    /** The button size. */
    size?: 'default' | 'large' | 'small' | undefined;
    /** If true, activates a button with a heavier font weight. */
    strong?: boolean;
    /** Specifies where to display the linked URL. Only applies when an href is provided. Special keywords: "_blank", "_self", "_parent", "_top". */
    target?: '_blank' | '_self' | '_parent' | '_top' | undefined;
    /** The type of the button. */
    type?: 'button' | 'submit' | 'reset';
}
// ...
export interface ValidatorType {
    validator: string;
    text: stringOrExpr;
    behavior?: FormBehaviorType[];
    class?: string;
}
/** Base definition for all component containers. */
export interface BaseComponentType extends BaseControlType {
    /**
     * Responsive size for the `ion-col` wrapping the component.
     * Can be a single value or a breakpoint map.
     *
     * ```typescript
     * size: 12
     * // or
     * size: { xs: 12, md: 6, lg: 'auto' }
     * ```
     *
     * ```html
     * <ion-col size="12"></ion-col>
     * <ion-col size-xs="12" size-md="6" size-lg="auto"></ion-col>
     * ```
     */
    size?: GridColSizeType | {
        xs?: GridColSizeType;
        sm?: GridColSizeType;
        md?: GridColSizeType;
        lg?: GridColSizeType;
        xl?: GridColSizeType;
    } | undefined;
    /**
     * Key used to evaluate user permissions via the `hasPermission` helper in your host.
     */
    permission?: string | boolean;
    /**
     * Expression to compute the disabled state of the component.
     *
     * ```typescript
     * component: { disabled: '!initialized || frm.invalid || frm.pristine' }
     * ```
     */
    disabled?: boolean | string;
    /**
     * Expression to compute the visibility of the component.
     *
     * ```typescript
     * component: { visible: 'frm.invalid || frm.pristine' }
     * ```
     */
    visible?: boolean | string;
    /**
     * Expression that evaluates to the form control validators.
     * ```typescript
     * component: { validators: '[Validators.required, Validators.email]' }
     * const validators = evalExpr(component.validators, { Validators }));
     * const control = new FormControl(defaultValue, validators)
     * ```
     */
    validators?: string;
    /**
     * Validation errors to display for the field.
     *
     * ```typescript
     * component: { errors: [{ validator: 'required', text: 'login.email_required' }] }
     * ```
     *
     * If `behavior` is not provided, all are considered (`FormBehaviorType`).
     * If `class` is not provided, 'error' is used by default.
     * ```typescript
     * const err: ValidatorType = {
     *   validator: 'invalid',
     *   text: 'generic.email_invalid',
     *   behavior: ['touched', 'dirty', 'isNew'],
     *   class: 'error'
     * }
     * ```
     */
    errors?: ValidatorType[];
    /**
     * Expression evaluated for the component's `click` event.
     *
     * ```typescript
     * component: { click: `saveRow()` }
     * ```
     *
     * ```html
     * <ion-button (click)="saveRow()">
     * ```
     *
     * @category Method
     */
    click?: string;
}
export interface ButtonComponentType extends BaseComponentType {
    type: 'button';
    icon?: string | IconControlType;
    label?: string | LabelControlType;
    button?: ButtonControlType;
}
export interface CheckboxComponentType extends BaseComponentType {
    type: 'checkbox';
    icon?: string | IconControlType;
    label?: string | LabelControlType;
    checkbox?: CheckboxControlType;
}
export interface DatetimeComponentType extends BaseComponentType {
    type: 'datetime';
    icon?: string | IconControlType;
    label?: string | LabelControlType;
    datetime?: DatetimeControlType;
}
export interface InputComponentType extends BaseComponentType {
    type: 'input';
    icon?: string | IconControlType;
    label?: string | LabelControlType;
    input?: InputControlType;
}
export interface ToggleComponentType extends BaseComponentType {
    type: 'toggle';
    icon?: string | IconControlType;
    label?: string | LabelControlType;
    toggle?: ToggleControlType;
}
export type FieldComponentType = ButtonComponentType | CheckboxComponentType | DatetimeComponentType | InputComponentType | ToggleComponentType;
export interface FieldColumnType {
    type: DataType;
    length?: number;
    decimal?: number;
    /** Field accepts nulls. @default true */
    nullable?: boolean;
    autoIncrement?: boolean;
    /** Indicates whether this field has unique values, such as keys. @default false */
    unique?: boolean;
    /** Indexed field. The name follows the pattern: `IDX_entityName_fieldName` */
    index?: boolean;
    virtual?: boolean;
    comment?: string;
    charset?: CharsetType;
    collation?: string;
    binnary?: boolean;
    /** Indicates whether values are hashed/encrypted. */
    hash?: boolean;
    /** Default value. If not provided, null is used. */
    default?: any;
}
export interface FieldType {
    /** Field name. Used with `formControlName` to bind the control in reactive forms. */
    name?: string;
    /** Declares the value type of the component for the front-end.
     *
     * ```typescript
     * value: 'boolean'
     * ```
     * To set a default value:
     * ```typescript
     * value: { type: 'boolean', default: true }
     * ```
     *
     * ```typescript
     * const defaultValue = field.value.default;
     * const control = new FormControl(defaultValue, validators)
     * ```
     */
    value?: ValueType | {
        type: ValueType;
        default?: any;
    } | undefined;
    column?: FieldColumnType;
    component?: FieldComponentType;
}